Please enter the commit message for your changes. Lines starting with '#' will be ignored, and an empty message aborts the commit.
On branch removebinary
Your branch is ahead of 'origin/removebinary' by 1 commit.
  (use "git push" to publish your local commits)

Changes to be committed:
	modified:   PirateInterpreter/Interpreters/ComparisonOperationInterpreter.cs
	modified:   PirateInterpreter/Interpreters/ForLoopStatementInterpreter.cs
	modified:   PirateInterpreter/Interpreters/ValueInterpreter.cs
	modified:   PirateInterpreter/SymbolTable.cs
	modified:   PirateInterpreter/Values/BaseValue.cs
	modified:   PirateInterpreter/Values/BooleanValue.cs
	modified:   PirateInterpreter/Values/StringValue.cs
	modified:   PirateLexer.Test/KeyWordServiceTest.cs
	modified:   PirateLexer/Interfaces/IKeyWordService.cs
	modified:   PirateLexer/TokenRepository.cs
	modified:   PirateLexer/Tokens/KeyWordService.cs
	modified:   PirateLexer/Tokens/Token.cs
	modified:   PirateParser/Node/VariableDeclarationNode.cs
	modified:   PirateParser/Parsers/ForLoopStatementParser.cs

+++ PirateInterpreter/Interpreters/ComparisonOperationInterpreter.cs
@@ -37,7 +37,7 @@ public class ComparisonOperationInterpreter : BaseInterpreter
             case TokenType.GREATERTHAN:
                 if ((left.Value is int || left.Value is Int64) && (right.Value is int || right.Value is Int64))
                 {
-                    if(Convert.ToInt32(left.Value) > Convert.ToInt32(right.Value))
+                    if(Convert.ToInt64(left.Value) > Convert.ToInt64(right.Value))
                     {
                         value = 1;
                     }
@@ -46,7 +46,7 @@ public class ComparisonOperationInterpreter : BaseInterpreter
             case TokenType.GREATERTHANEQUALS:
                 if ((left.Value is int || left.Value is Int64) && (right.Value is int || right.Value is Int64))
                 {
-                    if (Convert.ToInt32(left.Value) >= Convert.ToInt32(right.Value))
+                    if (Convert.ToInt64(left.Value) >= Convert.ToInt64(right.Value))
                     {
                         value = 1;
                     }
@@ -56,7 +56,7 @@ public class ComparisonOperationInterpreter : BaseInterpreter
             case TokenType.LESSTHAN:
                 if ((left.Value is int || left.Value is Int64) && (right.Value is int || right.Value is Int64))
                 {
-                    if (Convert.ToInt32(left.Value) < Convert.ToInt32(right.Value))
+                    if (Convert.ToInt64(left.Value) < Convert.ToInt64(right.Value))
                     {
                         value = 1;
                     }
@@ -65,7 +65,7 @@ public class ComparisonOperationInterpreter : BaseInterpreter
             case TokenType.LESSTHANEQUALS:
                 if ((left.Value is int || left.Value is Int64) && (right.Value is int || right.Value is Int64))
                 {
-                    if (Convert.ToInt32(left.Value) >= Convert.ToInt32(right.Value))
+                    if (Convert.ToInt64(left.Value) >= Convert.ToInt64(right.Value))
                     {
                         value = 1;
                     }
